[build] Fix implicit conversion from int64_t to uint64_t

+++ platform/default/src/mbgl/util/timer.cpp
@@ -60,8 +60,8 @@ Timer::Timer()
 Timer::~Timer() = default;
 
 void Timer::start(Duration timeout, Duration repeat, std::function<void()>&& cb) {
-    impl->start(std::chrono::duration_cast<Milliseconds>(timeout).count(),
-                std::chrono::duration_cast<Milliseconds>(repeat).count(),
+    impl->start(static_cast<uint64_t>(std::chrono::duration_cast<Milliseconds>(timeout).count()),
+                static_cast<uint64_t>(std::chrono::duration_cast<Milliseconds>(repeat).count()),
                 std::move(cb));
 }
